feat: 增加 URL 路径和权限之间的关系配置类

+++ push-app/src/main/java/com/dato/push/app/dao/SysRoleMenu.java
@@ -20,11 +20,11 @@ public class SysRoleMenu {
      * 角色ID
      */
     @Column("role_id")
-    private Long roleId;
+    private Integer roleId;
 
     /**
      * 菜单id
      */
     @Column("menu_id")
-    private Long menuId;
+    private Integer menuId;
 }
